requirements.txt
aiogram==3.19.0
colorlog==6.9.0
python-dotenv==1.1.0
aiohttp==3.9.3
